+++ src/js/pages/main/index.js
@@ -1,0 +1,15 @@
+import { h, Component } from 'preact';
+
+import Wall from '../wall';
+import Home from '../home';
+
+export default class Default extends Component {
+
+	constructor(props){
+		super(props)
+	}
+
+	render() {
+		return this.props.user ? <Wall {...this.props} /> : <Home {...this.props} />
+	}
+}