+++ CHANGELOG.md
@@ -1,2 +1,10 @@
+# private_beta: v0.1
 
-beta_v01: initialized reactjs, initialized firebase
+* blockstack auth: login/logout
+* websocket pub/sub to the orion8
+* wall feed loading
+* tag discover page
+* init settings page
+* search component in header
+* tag component (3 functions, follow, unfollow, go to the tag)
+* comments (can comment, comment list, hide some comments if more then 3)